% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_multi_stage_time_variant_2sex.R
\name{all_kin_dy}
\alias{all_kin_dy}
\title{Title time invariant two-sex multi-state kin projections}
\usage{
all_kin_dy(
  Uf,
  Um,
  Ff,
  Fm,
  alpha,
  na,
  ns,
  Parity,
  sex_Focal,
  Initial_stage_Focal
)
}
\arguments{
\item{Uf}{matrix (block structured). transfers female individuals across stages and advances their age (conditional on survial)}

\item{Um}{matrix (block structured). transfers male individuals across stages and advances their age (conditional on survial)}

\item{Ff}{matrix (block structured). accounts for female reproduction, and assigns newborns into given age*stage}

\item{Fm}{matrix (block structured). accounts for male reproduction; assigns newborns into age-class, and stage}

\item{alpha}{scalar. birth ratio (male:female)}

\item{na}{scalar. number of ages.}

\item{ns}{scalar. number of stages.}

\item{Parity}{logical. If true then we omit mothers of parity 0, and re-scale the mother's age*stage of parenting}

\item{sex_Focal}{logical. Female or Male}

\item{Initial_stage_Focal}{numeric. Any natural number {1,2,3,4,...}}
}
\value{
a list of matrices. Each list entry represents a particular kin. Each kin is chacacterised by a matrix of dimension:
nrow = 2* na * ns (2-sex age-stage structured) and ncol = na (Focal's age)
yielding the age*stage distribution of kin for each age of Focal
}
\description{
Title time invariant two-sex multi-state kin projections
}
