% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_multi_stage_time_variant_2sex.R
\name{create_full_dists_df}
\alias{create_full_dists_df}
\title{Title joint age*stage distributions of kin by each age of Focal, for each time period, and cohort of birth}
\usage{
create_full_dists_df(
  kin_matrix_lists,
  kin_names,
  years,
  start_year,
  na,
  ns,
  specific_kin
)
}
\arguments{
\item{kin_matrix_lists}{list of lists of kin matrices: list( list(X_focal), list(X_parents), ... ). Outer list is length 14  = number of kin. Inner lists have lenght = timescale
so list(X_focal) = list(X_focal\link{year1},X_focal\link{year2},...,X_focal\link{yearlast})}

\item{kin_names}{list of characters. Corresponding to above lists: list("F","m",....)}

\item{years}{vector. The timescale on which we implement the kinship model.}

\item{start_year}{. First year of varying vital rates (e.g., if years = 1990:2000 then start_year = 1990)}

\item{na}{numeric. Number of ages.}

\item{ns}{numeric. Number of stages.}

\item{specific_kin}{character. names of kin we wish to analyse, e.g., list("os","ys"). If null returns all 14.}
}
\value{
A data frame which gives for each age of Focal at each year in the timescale, the full age*stage dist of kin
}
\description{
Title joint age*stage distributions of kin by each age of Focal, for each time period, and cohort of birth
}
