% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.R
\name{dead_age_reasign}
\alias{dead_age_reasign}
\title{Reassign kin dead to proper Focal age}
\usage{
dead_age_reasign(kin)
}
\arguments{
\item{kin}{table. A kin output in table format from function \code{kin_time_invariant}, \code{kin_time_variant}, \code{kin_time_invariant_2sex}, \code{kin_time_invariant_2sex}, \code{kin_multi_stage}.}
}
\description{
Reassign death to proper Focal risk age
}
\details{
Matrix methods set dead experience in the next age where risk happened. This function return dead experienced in x from x+1.
}
