% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demokin_codes}
\alias{demokin_codes}
\title{DemoKin codes, Caswell (2020) codes, and useful labels.}
\format{
A data.frame with codes and labels for distinction between kin types.
}
\usage{
demokin_codes
}
\description{
DemoKin codes, Caswell (2020) codes, and useful labels.
}
\keyword{datasets}
