% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f_caswell_2021}
\alias{f_caswell_2021}
\title{Historic and projected fertility ratios from Sweden used in Caswell (2021)}
\format{
A matrix f with years as cols and ages as rows.
}
\source{
Caswell (2019)
}
\usage{
f_caswell_2021
}
\description{
Historic and projected fertility ratios from Sweden used in Caswell (2021)
}
\keyword{datasets}
