% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{F_mat_fem_edu}
\alias{F_mat_fem_edu}
\title{Singapore: Lists of matrices containing fertility rates by age and education. Females}
\format{
The data is aggregated into 5-year age groups and 5-year time intervals
}
\source{
Wittgenstein Center
}
\usage{
F_mat_fem_edu
}
\description{
Singapore: Lists of matrices containing fertility rates by age and education. Females
}
\keyword{datasets}
