% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Female_parity_mortality_list_UK}
\alias{Female_parity_mortality_list_UK}
\title{UK female parity mortality from 1965 to 2022}
\format{
list of age by stage matrices, entries give female parity mortality List starting 1965 ending 2022.
}
\source{
HFD and ONS
}
\usage{
Female_parity_mortality_list_UK
}
\description{
UK female parity mortality from 1965 to 2022
}
\keyword{datasets}
