% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fra_asfr_sex}
\alias{fra_asfr_sex}
\title{Fertility for France (2012) by sex in Caswell (2022).}
\format{
A data.frame with age specific fertility rates by age and sex.
}
\source{
Caswell (2022)
}
\usage{
fra_asfr_sex
}
\description{
Fertility for France (2012) by sex in Caswell (2022).
}
\keyword{datasets}
