% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fra_surv_sex}
\alias{fra_surv_sex}
\title{Survival probability for France (2012) by sex in Caswell (2022).}
\format{
A data.frame with survival probabilities by age and sex.
}
\source{
Caswell (2022)
}
\usage{
fra_surv_sex
}
\description{
Survival probability for France (2012) by sex in Caswell (2022).
}
\keyword{datasets}
