% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HMDHFD.R
\name{get_HMDHFD}
\alias{get_HMDHFD}
\title{Get time serie matrix data from HMD/HFD}
\usage{
get_HMDHFD(
  country = "SWE",
  min_year = 1900,
  max_year = 2018,
  user_HMD = NULL,
  pass_HMD = NULL,
  user_HFD = NULL,
  pass_HFD = NULL,
  OAG = 100
)
}
\arguments{
\item{country}{numeric. Country code from rom HMD/HFD.}

\item{min_year}{integer. Older year to get data.}

\item{max_year}{numeric. Latest year to get data.}

\item{user_HMD}{character. From HMD/HFD.}

\item{pass_HMD}{character. From HMD/HFD.}

\item{user_HFD}{character. From HMD/HFD.}

\item{pass_HFD}{character. From HMD/HFD.}

\item{user}{character. From HMD/HFD.}
}
\value{
A list wiith female survival probability, survival function, fertility and poopulation age specific matrixes, with calendar year as colnames.
}
\description{
Wrapper function to get data.frames of female survival, fertlity and population
of selected countriy on selected period.
}
