% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H_mat_edu}
\alias{H_mat_edu}
\title{Singapore: List of matrices that redistribute newborns to age-class 1 and "no education" category}
\format{
The data is aggregated into 5-year age groups and 5-year time intervals
}
\source{
Wittgenstein Center
}
\usage{
H_mat_edu
}
\description{
Singapore: List of matrices that redistribute newborns to age-class 1 and "no education" category
}
\keyword{datasets}
