% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_multi_stage_time_variant_2sex.R
\name{kin_multi_stage_time_variant_2sex}
\alias{kin_multi_stage_time_variant_2sex}
\title{Estimate kin counts by age, stage, and sex, in a time variant framework}
\usage{
kin_multi_stage_time_variant_2sex(
  U_list_females = NULL,
  U_list_males = NULL,
  F_list_females = NULL,
  F_list_males = NULL,
  T_list_females = NULL,
  T_list_males = NULL,
  H_list = NULL,
  birth_female = 0.49,
  parity = FALSE,
  output_kin = FALSE,
  summary_kin = TRUE,
  sex_Focal = "Female",
  initial_stage_Focal = NULL,
  output_years
)
}
\arguments{
\item{U_list_females}{list with matrix entries: period-specific female survival probabilities. Age in rows and states in columns.}

\item{U_list_males}{list with matrix entries: period-specific male survival probabilities. Age in rows and states in columns.}

\item{F_list_females}{list with matrix with elements: period-specific female fertility (age in rows and states in columns).}

\item{F_list_males}{list with matrix entries: period-specific male fertility (age in rows and states in columns).}

\item{T_list_females}{list of lists with matrix entries: each outer list entry is period-specific, and composed of
a list of stochastic matrices which describe age-specific female probabilities of transferring stage}

\item{T_list_males}{list of lists with matrix entries: each outer list entry is period-specific, and composed of
a list of stochastic matrices which describe age-specific male probabilities of transferring stage}

\item{H_list}{list with matrix entries: redistribution of newborns across each stage to a specific age-class}

\item{birth_female}{numeric. birth ratio of females to males in population}

\item{parity}{logical. parity states imply age distribution of mothers re-scaled to not have parity 0 when Focal born. Default \code{TRUE}.}

\item{output_kin}{vector. A vector of particular kin one wishes to obtain results for, e.g., c("m","d","oa"). Default is all kin types.}

\item{summary_kin}{logical. Results as a data frame of accumulated kin by age of Focal if TRUE, and kin by their age*stage distribution by age of Focal if FALSE.}

\item{sex_Focal}{character. Female or Male as the user requests.}

\item{initial_stage_Focal}{Numeric in Natural number set {1,2,...,}. The stage which Focal is born into (e.g., 1 for parity 0)}

\item{output_years}{vector. The times at which we wish to count kin: start year = output_years\link{1}, and end year = output_years\link{length.}}
}
\value{
A data frame with focal age, kin age, kin stage, kin sex, year, cohort, and expected number of kin given these restrictions.
}
\description{
Implementation of combined formal demographic models: Caswell II,III,IV.
}
