% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_non_stable.R
\name{kin_non_stable}
\alias{kin_non_stable}
\title{Estimate kin counts in a non stable framework}
\usage{
kin_non_stable(
  U = NULL,
  f = NULL,
  N = NULL,
  pi = NULL,
  focal_cohort = NULL,
  focal_year = NULL,
  selected_kin = NULL,
  birth_female = 1/2.04,
  Pb = FALSE
)
}
\arguments{
\item{U}{numeric. A matrix of survival ratios with rows as ages and columns as years. The name of each col must be the year.}

\item{f}{numeric. A matrix of age-specific fertility rates with rows as ages and columns as years. The name of each col must be the year.}

\item{N}{numeric. A matrix of population with rows as ages and columns as years. The name of each col must be the year.}

\item{pi}{numeric. A matrix with distribution of childbearing with rows as ages and columns as years. The name of each col must be the year.}

\item{focal_cohort}{integer. Year of birth of focal. Could be a vector. Should be within input data years range.}

\item{focal_year}{integer. Year of focal. Could be a vector. Should be within input data years range.}

\item{selected_kin}{character. kin to return: "m" for mother, "d" for daughter,...}

\item{birth_female}{numeric. Female portion at birth.}

\item{Pb}{logic. Is given Pb as the first row in P?. If not, takes \code{P(0,1)} as \code{P(b,1)}. Useful for fertility matrix first row. Default \code{FALSE}.}
}
\value{
A data frame with focal´s age, related ages and type of kin
(for example \code{d} is daughter, \code{oa} is older aunts, etc.), alive and death.
}
\description{
Implementation of non-stable Goodman-Keyfitz-Pullum equations based on Caswell (2021).
}
