% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kin_svk1990_caswell2020}
\alias{kin_svk1990_caswell2020}
\title{Output for Slovakia 1990 in Caswell (2020).}
\format{
A list with specific kin types age-stage matrix
}
\source{
Caswell (2021)
}
\usage{
kin_svk1990_caswell2020
}
\description{
Output for Slovakia 1990 in Caswell (2020).
}
\keyword{datasets}
