% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kins.R
\name{kins}
\alias{kins}
\title{Estimate kin counts}
\usage{
kins(
  ego_age = NULL,
  year = NULL,
  P = NULL,
  asfr = NULL,
  N = NULL,
  stable = FALSE,
  age = 0:100,
  birth_female = 1/2.04,
  alive = "yes"
)
}
\arguments{
\item{ego_age}{integer. Age where ego is.}

\item{year}{integer. Year where ego is with \code{ego_age} age.}

\item{P}{numeric. A matrix of survival ratios with rows as ages and columns as years.
The name of each col must be the year.}

\item{asfr}{numeric. A matrix of age-specific fertility rates with rows as ages and
columns as years. The name of each col must be the year.}

\item{N}{numeric. A matrix of population with rows as ages and columns as years.
The name of each col must be the year.}

\item{stable}{logical. Stable assumption given \code{year} rates.}

\item{age}{integer. Ages, assuming last one as an open age group.}

\item{birth_female}{numeric. Female portion at birth.}

\item{alive}{character. Only living kin counts \code{yes}, death kins \code{no}, or both other char.}
}
\value{
A list with:
\itemize{
\item A data frame with ego´s age \code{x}, related ages \code{x_kin} and kind of kin (for example \code{d} is daughter, \code{oa} is older aunts, etc.).
\item A data frame with available kins at each age of ego.
\item A data frame with available kins at actual age of ego.
\item Mean age of each type of kin at actual age of ego.
\item Total of each type of kin at actual age of ego.
}
}
\description{
Implementation of Goodman-Keyfitz-Pullum equations in an stable or not framework.
}
\details{
See Caswell (2019) for details on stable option.
}
\examples{
# If ego is 30 years old and lives in 1950. How much live kins would have if
# her relatives would experience mortality and fertility in that calendar year?
\dontrun{
swe30_1950_stable <- kins(ego_age = 30, year = 1950,
                          P = swe_surv, asfr = swe_asfr,
                          stable = TRUE)
# How much live kins would have if her relatives would experience
# mortality and fertility at each observed year?
swe30_1950_nonstable <- kins(ego_age = 30, year = 1950,
                          P = swe_surv, asfr = swe_asfr,
                          stable = FALSE)
# Difference in total by kin:
swe30_1950_stable[["kins_total"]] - swe30_1950_nonstable[["kins_total"]]
}

}
