% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kins_stable.R
\name{kins_stable}
\alias{kins_stable}
\title{Estimate kin counts in a stable framework}
\usage{
kins_stable(
  P = NULL,
  asfr = NULL,
  age = 0:(length(P) - 1),
  birth_female = 1/2.04,
  pi_stable = FALSE
)
}
\arguments{
\item{P}{numeric. A vector of survival ratios.}

\item{asfr}{numeric. A vector of age-specific fertility rates.}

\item{age}{integer. Ages, assuming last one as an open age group.}

\item{birth_female}{numeric. Female portion at birth.}

\item{pi_stable}{logical. Want mean age at childbearing as a result too. Default \code{FALSE}}
}
\value{
A data frame with ego´s age \code{x}, related ages \code{x_kin} and kind of kin
(for example \code{d} is daughter, \code{oa} is older aunts, etc.).
}
\description{
Implementation of Goodman-Keyfitz-Pullum equations adapted by Caswell (2019).
}
