% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Male_parity_mortality_list_UK}
\alias{Male_parity_mortality_list_UK}
\title{UK male parity mortality from 1965 to 2022}
\format{
list of age by stage matrices, entries give male parity mortality List starting 1965 ending 2022.
}
\source{
HFD and ONS
}
\usage{
Male_parity_mortality_list_UK
}
\description{
UK male parity mortality from 1965 to 2022
}
\keyword{datasets}
