% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_time_variant.R
\name{output_period_cohort_combination}
\alias{output_period_cohort_combination}
\title{APC combination to return}
\usage{
output_period_cohort_combination(
  output_cohort = NULL,
  output_period = NULL,
  age = NULL,
  years_data = NULL
)
}
\arguments{
\item{output_cohort}{integer. A vector with selected calendar years.}

\item{output_period}{integer. A vector with selected cohort years.}

\item{age}{integer. A vector with ages from the kinship network to be filtered.}

\item{years_data}{integer. A vector with years from the time-varying kinship network to be filtered.}
}
\value{
data.frame with years and ages to filter in \code{kin} and \code{kin_2sex} functions.
}
\description{
define APC combination to return in \code{kin} and \code{kin2sex}.
}
\details{
Because returning all period and cohort data from a huge time-series would be hard memory consuming,
this function is an auxiliary one to deal with selection from inputs \code{output_cohort} and \code{output_period}.
}
