% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Parity_transfers_by_age_list_UK}
\alias{Parity_transfers_by_age_list_UK}
\title{UK female parity transitions from 1965 to 2022}
\format{
list of age by stage matrices, entries give female parity transitions. List starting 1965 ending 2022.
}
\source{
HFD and ONS
}
\usage{
Parity_transfers_by_age_list_UK
}
\description{
UK female parity transitions from 1965 to 2022
}
\keyword{datasets}
