% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pi_caswell_2021}
\alias{pi_caswell_2021}
\title{Historic and projected mother´s age distribution of childbearing from Sweden used in Caswell (2021)}
\format{
A matrix pi with years as cols and ages as rows.
}
\source{
Caswell (2019)
}
\usage{
pi_caswell_2021
}
\description{
Historic and projected mother´s age distribution of childbearing from Sweden used in Caswell (2021)
}
\keyword{datasets}
