% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_multi_stage_time_variant_2sex.R
\name{pi_mix_TV}
\alias{pi_mix_TV}
\title{Mixing distributions for the time-variant multi-state 2-sex model: Non-parity case}
\usage{
pi_mix_TV(Ff, Fm, alpha, na, ns, previous_age_stage_dist)
}
\arguments{
\item{Ff}{matrix. Block-structured matrix which counts reproduction by females and assigns newborns an age and stage}

\item{Fm}{matrix. Block-structured matrix which counts reproduction by males and assigns newborns an age and stage}

\item{alpha}{scalar. Birth ratio male:female}

\item{na}{scalar. Number of age-classes}

\item{ns}{scalar. Number of stages}

\item{previous_age_stage_dist}{vector. Last years population structure (age\emph{stage}sex full distribution)}
}
\value{
list (of vectors). list[\link{1}] = full age\emph{stage}sex distribution
list[\link{2}] = female age\emph{stage distribution normalised
list[\link{3}] = male age}stage distribution normalised
list[\link{4}] = female marginal age distribution normalised
list[\link{5}] = male marginal age distribution normalised
}
\description{
Mixing distributions for the time-variant multi-state 2-sex model: Non-parity case
}
