% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagramm.R
\name{plot_diagram}
\alias{plot_diagram}
\title{plot a Kin diagram (network)}
\usage{
plot_diagram(kin_total, rounding = 3)
}
\arguments{
\item{kin_total}{data.frame. values in column \code{kin} define the relative type - see \code{demokin_codes()}. Values in column \code{count} are the expected number of relatives.}

\item{rounding}{numeric. Number of decimals to show in diagram.}
}
\value{
A Keyfitz-style kinship plot.
}
\description{
Draws a Keyfitz-style kinship diagram given a kinship object created by the \code{kin} function. Displays expected kin counts for a Focal aged 'a'.
}
