% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Redistribution_by_parity_list_UK}
\alias{Redistribution_by_parity_list_UK}
\title{UK parity assign parity at birth}
\format{
list of matrices which redistributes newborns to age-class 1 and parity 0. No time-variation.
}
\source{
None
}
\usage{
Redistribution_by_parity_list_UK
}
\description{
UK parity assign parity at birth
}
\keyword{datasets}
