% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svk_fxs}
\alias{svk_fxs}
\title{Female Slovakian fertility rates by parity stage in 1990 (Caswell, 2021)}
\format{
A matrix of fx with stages as cols and ages as rows.
}
\source{
Caswell (2021)
}
\usage{
svk_fxs
}
\description{
Female Slovakian fertility rates by parity stage in 1990 (Caswell, 2021)
}
\keyword{datasets}
