% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svk_Hxs}
\alias{svk_Hxs}
\title{Age where assign offspring of individuals in each partity stage (Caswell, 2021). All to zero age in this case.}
\format{
A matrix of ones in ages where assign offspring individuals, with stages as cols and ages as rows.
}
\source{
Caswell (2021)
}
\usage{
svk_Hxs
}
\description{
Age where assign offspring of individuals in each partity stage (Caswell, 2021). All to zero age in this case.
}
\keyword{datasets}
