% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svk_pxs}
\alias{svk_pxs}
\title{Female Slovakian survival probabilities by parity stage in 1990 (Caswell, 2021)}
\format{
A matrix of px with stages as cols and ages as rows.
}
\source{
Caswell (2021)
}
\usage{
svk_pxs
}
\description{
Female Slovakian survival probabilities by parity stage in 1990 (Caswell, 2021)
}
\keyword{datasets}
