% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svk_Uxs}
\alias{svk_Uxs}
\title{Probability of transition among parity stage for Slovakia in 1990, for each age, conditional on survival (Caswell, 2021).}
\format{
A list of column-stochastic matrix with probabilities of transition among parity stage, for each age, conditional on survival.
}
\source{
Caswell (2021)
}
\usage{
svk_Uxs
}
\description{
Probability of transition among parity stage for Slovakia in 1990, for each age, conditional on survival (Caswell, 2021).
}
\keyword{datasets}
