% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swe_pop}
\alias{swe_pop}
\title{Female swedish population from 1900 to 2015}
\format{
A matrix with years as cols and ages (0 to 100 as OAG) as rows.
}
\source{
HMD/HFD
}
\usage{
swe_pop
}
\description{
Female swedish population from 1900 to 2015
}
\keyword{datasets}
