% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T_mat_male_edu}
\alias{T_mat_male_edu}
\title{Singapore: Lists of transition matrices showing probabilities of moving between education states. Males}
\format{
The data is aggregated into 5-year age groups and 5-year time intervals
}
\source{
Wittgenstein Center
}
\usage{
T_mat_male_edu
}
\description{
Singapore: Lists of transition matrices showing probabilities of moving between education states. Males
}
\keyword{datasets}
