% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_time_variant_2sex.R
\name{timevarying_kin_2sex}
\alias{timevarying_kin_2sex}
\title{one time projection kin}
\usage{
timevarying_kin_2sex(Ut, Ft, Ft_star, pit, sex_focal, ages, pkin)
}
\arguments{
\item{Ut}{numeric. A matrix of survival probabilities (or ratios).}

\item{Ft}{numeric. A matrix of age-specific fertility rates.}

\item{Ft_star}{numeric. Ft but for female fertility.}

\item{pit}{numeric. A matrix with distribution of childbearing.
sex_focal}

\item{ages}{numeric.}

\item{pkin}{numeric. A list with kin count distribution in previous year.}
}
\description{
one time projection kin. internal function.
}
