% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{U_mat_fem_edu}
\alias{U_mat_fem_edu}
\title{Singapore: Lists of matrices containing survival probabilities by age (rows) and education (columns) from 2020-2090. Females}
\format{
The data is aggregated into 5-year age groups and 5-year time intervals
}
\source{
Wittgenstein Center
}
\usage{
U_mat_fem_edu
}
\description{
Singapore: Lists of matrices containing survival probabilities by age (rows) and education (columns) from 2020-2090. Females
}
\keyword{datasets}
