% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.R
\name{kin}
\alias{kin}
\title{Estimate kin counts}
\usage{
kin(
  U = NULL,
  f = NULL,
  time_invariant = TRUE,
  N = NULL,
  pi = NULL,
  output_cohort = NULL,
  output_period = NULL,
  output_kin = NULL,
  birth_female = 1/2.04
)
}
\arguments{
\item{U}{numeric. A vector or  matrix with probabilities (or survival ratios, or transition between age class in a more general perspective) with rows as ages (and columns as years in case of matrix, being the name of each col the year).}

\item{f}{numeric. Same as U but for fertility rates.}

\item{time_invariant}{logical. Constant assumption for a given \code{year} rates. Default \code{TRUE}.}

\item{N}{numeric. Same as U but for population distribution (counts or \verb{\%}). Optional.}

\item{pi}{numeric. Same as U but for childbearing distribution (sum to 1). Optional.}

\item{output_cohort}{integer. Vector of year cohorts for returning results. Should be within input data years range.}

\item{output_period}{integer. Vector of period years for returning results. Should be within input data years range.}

\item{output_kin}{character. kin types to return: "m" for mother, "d" for daughter,...}

\item{birth_female}{numeric. Female portion at birth.}
}
\value{
A list with:
\itemize{
\item \code{kin_full}: a data frame with year, cohort, focal´s age, related ages and type of kin (for example \code{d} is daughter, \code{oa} is older aunts, etc.), with living kin and death on that age.
\item \code{kin_summary}: a data frame with focal´s age, related ages and type of kin, with indicators obtained processing \code{kin_full}, grouping by cohort or period (depending arguments):
}
\itemize{
\item \code{count_living}: count of living kin at actual age of focal
\item \code{mean_age}: mean age of each type of living kin.
\item \code{sd_age}: standard deviation age of each type of living kin .
\item \code{count_death}: count of death kin at specific age of focal.
\item \code{count_cum_death}: cumulated count of death kin until specific age of focal.
\item \code{mean_age_lost}: mean age where focal lost her relative.
}
}
\description{
Implementation of Goodman-Keyfitz-Pullum equations in a matrix framework.
}
\details{
See Caswell (2019) and Caswell (2021) for details on formulas.
}
