% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_multi_stage.R
\name{kin_multi_stage}
\alias{kin_multi_stage}
\title{Estimate kin counts by age and stage in a time invariant framework}
\usage{
kin_multi_stage(
  U = NULL,
  f = NULL,
  D = NULL,
  H = NULL,
  birth_female = 1,
  output_kin = NULL,
  list_output = FALSE
)
}
\arguments{
\item{U}{list. age elemnts with column-stochastic transition matrix with dimension for the state space, conditional on survival.}

\item{f}{matrix. state-specific fertility (age in rows and states in columns).}

\item{D}{matrix. survival probabilities by state (age in rows and states in columns)}

\item{H}{matrix. assigns the offspring of individuals in some stage to the appropriate age class with 1 (age in rows and states in columns).}

\item{output_kin}{character. kin to return. For example "m" for mother, "d" for daughter. See the \code{vignette} for all kin types.}

\item{list_output}{logical. Results as a list. Default \code{FALSE}.}
}
\value{
A data frame with focal´s age, related ages and type of kin
(for example \code{d} is daughter, \code{oa} is older aunts, etc.), alive and death, and specific stage. If \code{list_output = TRUE} then this is a list.
}
\description{
Implementation of Goodman-Keyfitz-Pullum equations adapted by Caswell (2019). Stages are implied in length of input lists.
}
