% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_time_invariant.R
\name{kin_time_invariant}
\alias{kin_time_invariant}
\title{Estimate kin counts in a time invariant framework}
\usage{
kin_time_invariant(
  U = NULL,
  f = NULL,
  birth_female = 1/2.04,
  pi = NULL,
  output_kin = NULL,
  living = TRUE,
  list_output = FALSE
)
}
\arguments{
\item{U}{numeric. A vector of survival ratios with same length as ages.}

\item{f}{numeric. A vector of age-specific fertility rates with same length as ages.}

\item{birth_female}{numeric. Female portion at birth.}

\item{pi}{numeric. For using some specific non-stable age distribution of childbearing (same length as ages). Default \code{NULL}.}

\item{output_kin}{character. kin to return. For example "m" for mother, "d" for daughter. See the \code{vignette} for all kin types.}

\item{living}{logical. Only living kin counts \code{TRUE}, or including death kin \code{FALSE}.}

\item{list_output}{logical. Results as a list. Default \code{FALSE}}
}
\value{
A data frame with focal´s age, related ages and type of kin
(for example \code{d} is daughter, \code{oa} is older aunts, etc.), alive and death.
}
\description{
Implementation of Goodman-Keyfitz-Pullum equations adapted by Caswell (2019).
}
