% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_time_variant.R
\name{kin_time_variant}
\alias{kin_time_variant}
\title{Estimate kin counts in a time variant framework}
\usage{
kin_time_variant(
  U = NULL,
  f = NULL,
  N = NULL,
  pi = NULL,
  output_cohort = NULL,
  output_period = NULL,
  output_kin = NULL,
  birth_female = 1/2.04
)
}
\arguments{
\item{U}{numeric. A matrix of survival ratios with rows as ages and columns as years. Column names must be equal interval.}

\item{f}{numeric. A matrix of age-specific fertility rates with rows as ages and columns as years. Coincident with \code{U}.}

\item{N}{numeric. A matrix of population with rows as ages and columns as years. Coincident with \code{U}.}

\item{pi}{numeric. A matrix with distribution of childbearing with rows as ages and columns as years. Coincident with \code{U}.}

\item{output_cohort}{integer. Year of birth of focal to return as output. Could be a vector. Should be within input data years range.}

\item{output_period}{integer. Year for which to return kinship structure. Could be a vector. Should be within input data years range.}

\item{output_kin}{character. kin to return as output: "m" for mother, "d" for daughter,... See \code{vignette} for exahustive kin.}

\item{birth_female}{numeric. Female portion at birth.}
}
\value{
A data frame of population kinship structure, with focal's cohort, focal´s age, period year, type of relatives
(for example \code{d} is daughter, \code{oa} is older aunts, etc.), and age of (living or deceased) relatives.
}
\description{
Implementation of time variant Goodman-Keyfitz-Pullum equations based on Caswell (2021).
}
