% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kins.R
\name{kins}
\alias{kins}
\title{Estimate kin counts}
\usage{
kins(
  U = NULL,
  f = NULL,
  N = NULL,
  pi = NULL,
  stable = TRUE,
  ego_cohort = NULL,
  ego_year = NULL,
  selected_kins = NULL,
  birth_female = 1/2.04,
  Pb = FALSE,
  living = TRUE
)
}
\arguments{
\item{U}{numeric. A matrix of survival ratios with rows as ages and columns as years. The name of each col must be the year.}

\item{f}{numeric. A matrix of age-specific fertility rates with rows as ages and columns as years. The name of each col must be the year.}

\item{N}{numeric. A matrix of population with rows as ages and columns as years. The name of each col must be the year.}

\item{pi}{numeric. A matrix with distribution of childbearing with rows as ages and columns as years. The name of each col must be the year.}

\item{stable}{logical. Stable assumption given \code{year} rates, taking ego_year in case is U and f are matrix.}

\item{ego_cohort}{integer. Year of birth of ego. Could be a vector. Should be within input data years range.}

\item{ego_year}{integer. Year of ego. Could be a vector. Should be within input data years range.}

\item{selected_kins}{character. Kins to return: "m" for mother, "d" for daughter,...}

\item{birth_female}{numeric. Female portion at birth.}

\item{Pb}{logic. Is given Pb as the first row in P?. If not, takes \code{P(0,1)} as \code{P(b,1)}. Useful for fertility matrix first row. Default \code{FALSE}.}

\item{living}{logial. Only living kin counts \code{TRUE}, or including death kins \code{FALSE}.}
}
\value{
A list with:
\itemize{
\item \code{kins}: a data frame with ego´s age \code{x}, related ages \code{x_kin} and type of kin (for example \code{d} is daughter, \code{oa} is older aunts, etc.). The content (living or deaths) dependes on \code{alive} param.
\item A data frame with kins at each age of ego´s life. The content (living or deaths) dependes on \code{alive} param.
\item A data frame with kins at actual age of ego. The content (living or deaths) dependes on \code{alive} param.
\item Mean age of each type of kin at actual age of ego.  The content (living or deaths) dependes on \code{alive} param.
\item Total of each type of kin at actual age of ego.  The content (living or deaths) dependes on \code{alive} param.
}
}
\description{
Implementation of Goodman-Keyfitz-Pullum equations in a stable or not framework.
}
\details{
See Caswell (2019) and Caswell (2021) for details on formulas.
}
\examples{
# If ego is 30 years old and lives in 1950. How much live kins would have if
# her relatives would experience mortality and fertility in that calendar year?
\dontrun{
swe30_1950_stable <- kins(U = swe_surv, f = swe_asfr, ego_year = 1950, stable = TRUE,selected_kins = c("m","gm"))
# How much live mothers and grandmothers would have if her relatives would experience
# mortality and fertility at each observed year?
swe30_1950_nonstable <- kins(U = swe_surv, f = swe_asfr,N = swe_pop,
                             stable = FALSE, ego_year = 1950, selected_kins = c("m","gm"))
# Difference in total by kin:
swe30_1950_stable$kins_by_age_ego \%>\% filter(age_ego==30) \%>\% select(kin, total)
swe30_1950_nonstable$kins_by_age_ego \%>\% filter(age_ego==30) \%>\% select(kin, total)
}

}
