% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kins_non_stable.R
\name{kins_non_stable}
\alias{kins_non_stable}
\title{Estimate kin counts in a non stable framework}
\usage{
kins_non_stable(
  ego_age = NULL,
  year = NULL,
  P = NULL,
  asfr = NULL,
  N = NULL,
  age = 0:100,
  birth_female = 1/2.04,
  Pb = FALSE
)
}
\arguments{
\item{ego_age}{integer. Age where ego is.}

\item{year}{integer. Year where ego is with \code{ego_age} age.}

\item{P}{numeric. A matrix of survival ratios with rows as ages and columns as years. The name of each col must be the year.}

\item{asfr}{numeric. A matrix of age-specific fertility rates with rows as ages and columns as years. The name of each col must be the year.}

\item{N}{numeric. A matrix of population with rows as ages and columns as years. The name of each col must be the year.}

\item{age}{integer. Ages, assuming last one as an open age group.}

\item{birth_female}{numeric. Female portion at birth.}

\item{Pb}{logic. Is given Pb as the first row in P?. Default \code{FALSE}.}
}
\value{
A data frame with ego´s age \code{x}, related ages \code{x_kin} and kind of kin
(for example \code{d} is daughter, \code{oa} is older aunts, etc.).
}
\description{
Implementation of Goodman-Keyfitz-Pullum equations but as a
weigthed average of possible ages at childbear of mothers and grandmothers of ego.
}
