% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kins_stable.R
\name{kins_stable}
\alias{kins_stable}
\title{Estimate kin counts in a stable framework}
\usage{
kins_stable(
  U = NULL,
  f = NULL,
  birth_female = 1/2.04,
  pi = NULL,
  selected_kins = NULL,
  pi_stable = FALSE,
  list_output = FALSE
)
}
\arguments{
\item{U}{numeric. A vector of survival ratios.}

\item{f}{numeric. A vector of age-specific fertility rates.}

\item{birth_female}{numeric. Female portion at birth.}

\item{pi}{numeric. For using the non-stable distribution of childbearing. Default \code{NULL}.}

\item{selected_kins}{character. Kins to return: "m" for mother, "d" for daughter,...}

\item{pi_stable}{logical. Want mean age at childbearing as a result too. Default \code{FALSE}}

\item{list_output}{logical. Want kin results as a list. Default \code{FALSE}}
}
\value{
A data frame with ego´s age, related ages and type of kin
(for example \code{d} is daughter, \code{oa} is older aunts, etc.), alive and death.
}
\description{
Implementation of Goodman-Keyfitz-Pullum equations adapted by Caswell (2019).
}
