% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kins_stable.R
\name{kins_stable}
\alias{kins_stable}
\title{Estimate kin counts in a stable framework}
\usage{
kins_stable(
  p = NULL,
  f = NULL,
  age = 0:(length(p) - 1),
  cum_deaths = FALSE,
  birth_female = 1/2.04,
  pi = FALSE
)
}
\arguments{
\item{p}{numeric. A vector of survival ratios.}

\item{f}{numeric. A vector of age-specific fertility rates.}

\item{age}{integer. Ages, assuming last one as an open age group.}

\item{cum_deaths}{logic. Include or not expected lost. Default \code{FALSE}. See Caswell (2019) for more details.}

\item{birth_female}{numeric. Female portion at birth.}

\item{pi}{numeric. Want mean age at childbearing as result.}
}
\value{
A data frame with ego´s age \code{x}, related ages \code{x_kin} and kind of kin
(for example \code{d} is daughter, \code{oa} is older aunts, etc.).
}
\description{
Implementation of Goodman-Keyfitz-Pullum equations adapted by Caswell (2019).
}
