% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_time_variant.R, R/kin_time_variant_2sex.R
\name{output_period_cohort_combination}
\alias{output_period_cohort_combination}
\title{defince apc combination to return}
\usage{
output_period_cohort_combination(
  output_cohort = NULL,
  output_period = NULL,
  age = NULL,
  years_data = NULL
)

output_period_cohort_combination(
  output_cohort = NULL,
  output_period = NULL,
  age = NULL,
  years_data = NULL
)
}
\description{
defince apc to return.

defince apc to return.
}
