% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagramm.R
\name{plot_diagram}
\alias{plot_diagram}
\title{plot a Kin diagram (network)}
\usage{
plot_diagram(kins_total, rounding = 3)
}
\arguments{
\item{kins_total}{vector. Colum names must be as \code{kins} output.}

\item{rounding}{numeric. Estimation could have a lot of decimals. Rounding will make looks more clear the diagramm.}

\item{ego_age}{integer.}
}
\value{
A plot
}
\description{
Given estimation of kin counts from \code{kins} function, draw a diagramm.
}
