% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_time_variant.R
\name{timevarying_kin}
\alias{timevarying_kin}
\title{one time projection kin}
\usage{
timevarying_kin(Ut, ft, pit, ages, pkin)
}
\arguments{
\item{Ut}{numeric. A matrix of survival probabilities (or ratios).}

\item{ft}{numeric. A matrix of age-specific fertility rates.}

\item{pit}{numeric. A matrix with distribution of childbearing.}

\item{ages}{numeric.}

\item{pkin}{numeric. A list with kin count distribution in previous year.}
}
\value{
A list of 14 types of kin matrices (kin age by Focal age) projected one time interval.
}
\description{
one time projection kin. internal function.
}
